Move test support into common directory.
See here: https://doc.rust-lang.org/book/ch11-03-test-organization.html#submodules-in-integration-tests
